Fix compiler warnings

+++ FMDownsampler.h
@@ -36,7 +36,7 @@ private:
   CRingBuffer<TSamplePairPack> m_ringBuffer;
 
   uint32_t m_samplePack;
-  uint32_t *m_samplePackPointer;
+  uint8_t *m_samplePackPointer;
 
   uint8_t m_packIndex;
   uint8_t m_downSampleIndex;
